Up case now working, but should place the elevator case by hand and always at the left side of the upcase !

+++ Game/VS_Project/TerraSkweek/TerraSkweek/MobSpawner.cpp
@@ -63,7 +63,7 @@ bool MobSpawner::Draw()
 
 }
 
-MobSpawner::MobSpawner(int x, int y, string directory) : SpecialCase(x, y, directory)
+MobSpawner::MobSpawner(float x, float y, string directory) : SpecialCase(x, y, directory)
 {
 	animated = false;
